Implemented a rough blast attack with knockback.
A few notes here, the blast is using a sphere cast with random directions within a provided angle. Ideally, I prefer this to be a more uniform cone with a single sphere cast into the middle, with some sphere casts surrounding it in a 360 degree coverage. This will be a planned implementation.