Forgot to commit. A lot of updates.
Had to delete project in GitHub and create a new repo since I accidently added in a background music asset that was not made by me and I didn't know how to delete it. To make it easier if I add in other audio not made by me, I just added the whole audio folder to the git ignore.